all os: RT#106564 - prevent buffer overflow in format string
Thanks to Thomas Klausner for pointing this out.

Details:
Compiling with SSP shows that there is a buffer overflow when using
strcpy to copy Defaultformat into format, because it is too small.
F_LASTFIELD is 18 and 18 is the last used field, but the count starts
from 0, so we need F_LASTFIELD+2 spaces in format (one extra for the
terminating NUL).