Skip LXC hidden dir in /dev when looking for TTYs
In LXC containers, /dev/.lxc contains subdirs for the sys and proc fs to be
mounted into the container. These files are not readable directly,
Proc::ProcessTable::_get_tty_list attempts to stat the files in these
directories which results in a lot of warnings. There aren't going to be any TTY
device files there so skip the whole dir.